sof/sst: coexistence-based options
now that we have coexistence, keep HDaudio legacy and have single
option with SST/SOF options selected.

Baytrail is based on SOF, Haswell/Broadwell remain with SST for now.

For SKL/KBL, the SST driver is only used for Chromebooks so disable
the HDAudio codec support (which never quite worked)

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>